Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

Currently, the active repository for liberty is at https://github.com/attn1/htc-kernel-liberty .  No pull requests for these changes are active for that repo, and pre-built kernels & modules are not acceptable. 

I need to be able to build it, test it, and then we can merge it.

Patch-set: 2