Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

So lets fix the experimental themes or fix the backuptool. This change has nothing to do with why the backuptool does not run correctly with experimental themes. As of now, not building recovery results in bootloops unless system is formatted prior to flashing on upgrade.

Patch-set: 1