Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Ladios I do appreciate your work.  But can you submit your kernel patches to gerrit for review and not a precompiled kernel.  It is up to the device maintainer to upload precompiled kernels for the said device.  Which currently you are not.  If you upload the patches from your kernel source here we will review accordingly.  Thanks for your work though.

Patch-set: 1