Update patch set 4
Patch Set 4: I would prefer that you didn't submit this

Rebased. This will turn off the light which is now lights up by init.rc. This is not tested yet, but it should work. There's another option that is to add an one shot service that performs tasks in offmode, so recovery would boot faster.

Patch-set: 4